Extract test slot row mapper
